fix: import srp text color in dark mode and move validation logic to another file (#14718)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

This pr fixes the input text color to show the alternative text color in
dark mode and moves validation logic into a validation file.

Changes:
1. ImportNewSecretRecoveryPhrase dark mode text color.
2. Moved logic into a new validation file

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
3. What is the improvement/solution?
-->

Fixes:
https://consensyssoftware.atlassian.net/browse/MMMULTISRP-172?atlOrigin=eyJpIjoiODc2NzY1Y2RjMGExNDBiMzlkODY2OWFkOWNkYjIxMTUiLCJwIjoiaiJ9

1. Go to account actions
2. Click on `Secret Recovery Phrase`
4. Enter a word
5. Switch to dark mode and see that the text color is visible.

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

<!-- [screenshots/recordings] -->

![Screenshot 2025-04-17 at 19 13
05](https://github.com/user-attachments/assets/8dc35491-b45f-4eb5-8ceb-02ece09208f0)

<!-- [screenshots/recordings] -->

- [x] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.
