+++ app/components/Views/ImportNewSecretRecoveryPhrase/index.tsx
@@ -33,7 +33,6 @@ import Icon, {
 } from '../../../component-library/components/Icons/Icon';
 import ButtonLink from '../../../component-library/components/Buttons/Button/variants/ButtonLink';
 import { ButtonSize } from '../../../component-library/components/Buttons/Button';
-import { wordlist } from '@metamask/scure-bip39/dist/wordlists/english';
 import { isValidMnemonic } from '../../../util/validators';
 import useCopyClipboard from '../Notifications/Details/hooks/useCopyClipboard';
 import ClipboardManager from '../../../core/ClipboardManager';
@@ -46,6 +45,14 @@ import {
 } from '../../../component-library/components/Toast';
 import { useSelector } from 'react-redux';
 import { selectHDKeyrings } from '../../../selectors/keyringController';
+import {
+  validateSRP,
+  validateCompleteness,
+  validateCase,
+  validateWords,
+  validateMnemonic,
+} from './validation';
+import { AppThemeKey } from '../../../util/theme/models';
 
 const defaultNumberOfWords = 12;
 
@@ -105,9 +112,6 @@ const ImportNewSecretRecoveryPhrase = () => {
     }
   }, [inputWidth]);
 
-  const hasUpperCase = (draftSrp: string) =>
-    draftSrp !== draftSrp.toLowerCase();
-
   const hasEmptySrp = useMemo(
     () => secretRecoveryPhrase.every((word) => word === ''),
     [secretRecoveryPhrase],
@@ -135,110 +139,6 @@ const ImportNewSecretRecoveryPhrase = () => {
 
   const onSrpChange = useCallback(
     (newDraftSrp: string[]) => {
-      const validateSRP = (phrase: string[], words: boolean[]) => {
-        if (!phrase.some((word) => word !== '')) {
-          return { error: '', words };
-        }
-
-        const state = {
-          error: '',
-          words: phrase.map((word) => !wordlist.includes(word)),
-        };
-
-        return state;
-      };
-
-      const validateCompleteness = (
-        state: { error: string; words: boolean[] },
-        phrase: string[],
-      ) => {
-        if (state.error) {
-          return state;
-        }
-        if (phrase.some((word) => word === '')) {
-          return {
-            ...state,
-            error: strings(
-              'import_new_secret_recovery_phrase.error_number_of_words_error_message',
-            ),
-          };
-        }
-        return state;
-      };
-
-      const validateCase = (
-        state: { error: string; words: boolean[] },
-        phrase: string,
-      ) => {
-        if (state.error) {
-          return state;
-        }
-        if (hasUpperCase(phrase)) {
-          return {
-            ...state,
-            error: strings(
-              'import_new_secret_recovery_phrase.error_srp_is_case_sensitive',
-            ),
-          };
-        }
-        return state;
-      };
-
-      const validateWords = (state: { error: string; words: boolean[] }) => {
-        if (state.error) {
-          return state;
-        }
-
-        const invalidWordIndices = state.words
-          .map((invalid, index) => (invalid ? index + 1 : 0))
-          .filter((index) => index !== 0);
-
-        if (invalidWordIndices.length === 0) {
-          return state;
-        }
-        if (invalidWordIndices.length === 1) {
-          return {
-            ...state,
-            error: `${strings(
-              'import_new_secret_recovery_phrase.error_srp_word_error_1',
-            )}${invalidWordIndices[0]}${strings(
-              'import_new_secret_recovery_phrase.error_srp_word_error_2',
-            )}`,
-          };
-        }
-
-        const lastIndex = invalidWordIndices.pop();
-        const firstPart = invalidWordIndices.join(', ');
-        return {
-          ...state,
-          error: `${strings(
-            'import_new_secret_recovery_phrase.error_multiple_srp_word_error_1',
-          )}${firstPart}${strings(
-            'import_new_secret_recovery_phrase.error_multiple_srp_word_error_2',
-          )}${lastIndex}${strings(
-            'import_new_secret_recovery_phrase.error_multiple_srp_word_error_3',
-          )}`,
-        };
-      };
-
-      const validateMnemonic = (
-        state: { error: string; words: boolean[] },
-        phrase: string,
-      ) => {
-        if (state.error) {
-          return state;
-        }
-        if (!isValidMnemonic(phrase)) {
-          return {
-            ...state,
-            error: strings(
-              'import_new_secret_recovery_phrase.error_invalid_srp',
-            ),
-          };
-        }
-        return state;
-      };
-
       const hideErrorIfSrpIsEmpty = (
         state: { error: string; words: boolean[] },
         phrase: string[],
@@ -401,6 +301,10 @@ const ImportNewSecretRecoveryPhrase = () => {
                     borderColor: invalidSRPWords[index]
                       ? colors.error.default
                       : colors.border.muted,
+                    color:
+                      themeAppearance === AppThemeKey.light
+                        ? colors.text.default
+                        : colors.text.alternative,
                   }}
                   autoCapitalize="none"
                   keyboardAppearance={themeAppearance}
